Translated database content (used: KnpBehaviorsBundle + A2lixTranslationFormBundle)
