+++ app/config/config.yml
@@ -52,10 +52,10 @@ doctrine:
         dbname:   "%database_name%"
         user:     "%database_user%"
         password: "%database_password%"
-        charset:  utf8mb4
+        charset:  utf8
         default_table_options:
-            charset: utf8mb4
-            collate: utf8mb4_unicode_ci
+            charset: utf8
+            collate: utf8_general_ci
         # if using pdo_sqlite as your database driver:
         #   1. add the path in parameters.yml
         #     e.g. database_path: "%kernel.root_dir%/../var/data/data.sqlite"
@@ -82,6 +82,16 @@ stof_doctrine_extensions:
     orm:
         default:
             timestampable: true
+
+knp_doctrine_behaviors:
+    translatable: true
+
+a2lix_translation_form:
+    locale_provider: default
+    locales: [en, uk, ru, fr, es]
+    default_locale: en
+    required_locales: [en]
+    templating: "A2lixTranslationFormBundle::default.html.twig"
 
 knp_paginator:
     page_range: 5                      # default page range used in pagination control
