+++ composer.json
@@ -29,7 +29,9 @@
         "sensio/framework-extra-bundle": "^3.0.2",
         "incenteev/composer-parameter-handler": "^2.0",
         "stof/doctrine-extensions-bundle": "^1.2",
-        "knplabs/knp-paginator-bundle": "^2.5"
+        "knplabs/knp-paginator-bundle": "^2.5",
+        "knplabs/doctrine-behaviors": "~1.1",
+        "a2lix/translation-form-bundle": "^2.1"
     },
     "require-dev": {
         "sensio/generator-bundle": "^3.0",
